requirements.txt
altgraph==0.10.2
appdirs==1.4.4
asgiref==3.5.2
asn1crypto==0.24.0
bdist-mpkg==0.5.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
configparser==4.0.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.0
Django==4.1.2
django-cors-headers==3.13.0
django-pdfkit==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
enum34==1.1.6
filelock==3.0.12
future==0.17.1
idna==3.4
importlib-metadata==1.6.1
importlib-resources==2.0.1
inflection==0.5.1
ipaddress==1.0.22
itypes==1.2.0
Jinja2==3.1.2
macholib==1.5.1
MarkupSafe==2.1.1
modeltranslation==0.25
modulegraph==0.10.4
oauthlib==3.2.2
path==16.5.0
pathlib2==2.3.5
pdfkit==0.6.0
Pillow==9.3.0
py2app==0.7.3
pycparser==2.19
PyJWT==2.6.0
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==1.5
python3-openid==3.2.0
pytz==2013.7
requests==2.28.1
requests-oauthlib==1.3.1
scandir==1.10.0
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing==3.7.4.1
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.0.23
zipp==1.2.0
